Added komitee routes for displaying the individual komitees #5
