+++ smvsite/komitee/models.py
@@ -1,3 +1,10 @@
 from django.db import models
 
-# Create your models here.
+
+class Komitee(models.Model):
+    title = models.CharField(max_length=100, unique=True)
+    description = models.CharField(max_length=200)
+    body = models.TextField()
+
+    def __str__(self):
+        return self.title
