+++ smvsite/komitee/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path('<str:article_name>', views.blogEntry, name='index'),
+]
