+++ smvsite/landing/admin.py
@@ -1,4 +1,1 @@
-from django.contrib import admin
-from .models import Komitee
-
-admin.site.register(Komitee)
+# Register your admin site herer
